requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
idna==3.10
pillow==11.0.0
requests==2.32.3
urllib3==2.2.3
